Refactor tag distribution formatting into separate module
- Moved the `format_tag_distribution` method and related helper methods from `lib/timet/formatter.rb` to a new `TagDistribution` module in `lib/timet/tag_distribution.rb`.
- Renamed `format_tag_distribution` to `tag_distribution` for clarity.
- Updated `TimeReport` class in `lib/timet/time_report.rb` to include the `TagDistribution` module and call the new `tag_distribution` method.

This refactor improves code organization by separating concerns and makes the `Formatter` module more focused on general formatting tasks.
